Fix TableIterator::operator==
This fixes 2 bugs. First the bug I added myself in aa7841ef3f831ff60fd035179a9972dac15d679f (it was supposed
to be if (row != other.row) return false;) and second the if-condition which checked for table-columns which
had an additional ; at the end and therefore we always assumed they are not equal.