constify KoDocumentRdf
There are some ugly casts now in some places, but those at least mean
we can grep where the documentrdf or its model is changed. Next step
will be figuring out how to re-introduce shared pointers: the
Soprano::Model is passed around so widely that it becomes a bit scary.