Make sure the correct destructor is called
When you have a base class you should have a virtual destructor otherwise
the destructor if the inherited class is not called when you have a pointer
to the base class.
This fixes the problem that pattern which had a temporary file for their
image to delete the temporary file when the shape is destructed. The problem
happend for pages that had a image as page background in stage.