Remove viewconverter variable from KWViewMode
KWViewMode owned a pointer to a viewconverter. It used to be the viewconverter
taken from the canvas, then it was the viewconverter passed on construction.
By not making the viewconverter part of the state of KWViewMode, it's possible
to use the calculation methods with different viewconverter, for instance
for different zoomlevels, independent of the zoomlevel of the actual canvas.