Fix cache to not need to update all
When using QRegion it merges all rects into one rect which means the next
time the cache needs to be updated it will repaint all rending rectangles
in one go which can be very expensive at a high zoom level.
With the change the rects are kept and only the rects that are visible gets
repainted. That speeds up the cache.