Quick fix to make libkowmf build on windows.
Since there used to be a WmfStruct.h and a wmfstruct.h (obsolete), and Windows is not case sensitive, we have to separate the names.

In the long run, I will make sure that everything in the two unused
parsers is already used in the real WMF library and remove the other
ones.