WMF: Introduce a struct for the Playback Device Context.
This will allow us to make the parser more efficient, and in the
future also better support the quite complex transformations that the
Window/Viewport system allows.  Right now, these calculations are only
done in the painter backend which makes e.g. the karbon wmf filter
very wrong for some images.

The device context is only initialized right now, not used; that will
be the next commit.