Make Words handle documents with variable page sizes sensibly
When in zoom-to-page or zoom-to-page-width mode we want to see all pages
fully or to their full width. So cache the largest pagesize we've encountered
and use that to set the zoomcontroller's page size.

This also fixes the existential uncertainty of words what its page size
is for zooming purposes when it needs to show two differently sized pages
at the same time.
BUG:260235