Abuse the bmp filter for some other simple formats
BMP, GIF, XPM, XBM are file formats we don't want to spend much time on.
Krita is not for such files. So abuse the Qt support for those file formats
like we used to abuse GraphicsMagick or ImageMagick and that's it -- if
Qt cannot open these files, it's better to convert them using another
application before importing in Krita.

This also removes my gif filer which never worked really well.
BUG: 167713
BUG: 270634
BUG: 263987