improve updating size and position after (re)creating outline
before this change, we tried to keep the top left corner of our
text shape at the same position. however that is only correct as
long as they height of the text does not change, otherwise our
baseline would move in y-direction. Now we try harder to keep the
text baseline at the same position in document coordinates.