Don't read styles 3 times
Reuse the KoShapeLoadingContext for loading of headers and footers. With
this change it is no longer needed to parse the styles again when loading
the headers/footers as it reuses what is already there.
This should speedup loading of text documents having a lot of styles as the
styles loading will be done in a 1/3 of the time then before.