Add a Delete button that hovers over the lower right corner of a preset when selected
This update adds a new button that appears at the bottom right corner of a preset icon
when it is selected. The button is flat, has a trashcan icon, and when pressed,
deletes the currently selected preset. The button disappears when a new paintop is
selected, when the preset strip is scrolled with the side buttons, and when the paintop
settings popup appears.

The creation of this button required a small modification to KoResourceItemChooser:
slotButtonClicked and the enum of buttons were made public.