Improve xls progress reporting
If there where a lot of embedded object there was not progress update for quite some time.
Also fot the odf loading part there was no update as the sidewinder updates reached 100%
before.
This commit changes the wight of the different parts of loading and adds reporting to loading
of embedded objects/shapes.