Factor out common parts of BentConnector{2,3}
The logic for drawing bent connectors is VERY common. Exploit that to
simplify the code.

Factor out common parts of BentConnector{2,3}

The logic for drawing bent connectors is VERY common. Exploit that to
simplify the code. Lukas: if you like this, I think there is more
scope for this kind of thing.