Fix bug 275340 -  Crash occurred while opening ...
Fixed bug 275340 -  Crash occurred while opening a file containing picture in its footer
Problem was that the placeholder loaded the image but then the placeholder was not used
and destroyed which also destroyed the image it had loaded. However this image was still
in the data for loading and therefore the image was no longer valid when it was then loaded
as part of the text shape which resulted in a crash.
This is fixed by loading the attributes only when they are used.