WMF: Send the text rotation via the device context instead of as a parameter.
This is the first of a series of commits that will make more and more
information be sent to the backend via the playback device context.
This will simplify the API to the backend and make it simpler to
create new backends.

It will also remove some bugs that exist because of a confusion of
when a value is used.  This is such an example, since the text
rotation that was used was the last one that was defined definition
and put into the object stack (mTextRotation from
META_CREATEFONTINDIRECT in the parser) instead of when it was last
used (META_SELECTOBJECT).