Fix build of kdchart on opensolaris
Patch by tropichajma. Thank you for your patch!

"this is because the definition of on_sbStartColumn_valueChanged and
on_sbStartRow_valueChanged in KDChartDatasetSelector.cpp do not match
KDChartDatasetSelector.h - they differ in the const."

Note that when we re-import kdchart, we might have re-apply the patch
BUG:244050
CCMAIL:tropikhajma <tropikhajma@gmail.com>