Fix bug when there are more qtextdocuments
There was the posibility that the current item and and highlighted item
where not the same when the order of the qtextdocuments searched where not
the same as in the QHash and the possibility to hit that is quite big.
SO use the current match to find the correct entry in the QHash to update.
This might be a bit slower but still fast enough.