Use 32bits/pixel thumbnail preview.
Reducing to 8bpp reduces file sizes quite a lot, but we don't care nowdays we have a lot of hard disk, and high-speed internet, and anyway it is mostly krita which still use that function and the preview is an epsilon of the document size, and artists like to get good quality preview of their awesome work.

And anyway. The oasis thumbnail is 32bits, so lets bring the old format to the modern age.