Modify kis_dlg_layer_properties to show a preview of changes to the user
o Modify KisLayerManager and KisDlgLayerProperties to let the latter take charge of handling the changes to the layer.
o Add a ticker to keep change of changes to the configuration and show preview changes.
o Make the dialog non-modal.
o Modify wdglayerproperties.ui to give more room for widgets.
o ...Also add a preview CheckBox.

BUG:232141