Remove unnecessary clutter from the KisLayerBox menu
o Commands to change the layer's individual properties were removed: there are
icons for that.
o Commands to add new layers were removed: the add layer button is there for that,
and there is no need to right click on a layer to add a new layer.
o Commands to delete, duplicate or merge the layer down were presented: there are
icons to do that (except merge layer down), however, right clicking on a layer to
remove it or duplicate it makes sense, unlike right clicking on a layer to add a new one.

BUG:275764