Added a stroke strategy for processing QUndoCommand's
Now the commands can be added to the common queue and be run
in parallel using common threads. Such strokes can be canceled
as well.

And please do not look at AllThoseLongLongDisastrousClassNames.
This is going to be fixed after we decide what to do with
the naming rules. The best solution is namespaces, I guess.