Solve some Bit Rot in the KisPresetSelector strip
Natural margins have been added to KisPresetChooser or some class on which
KisPresetSelectorStrip depends, shrinking the inner area of the strip and
making the trash can appear in the wrong place.
Commit that (apparently as far as my testing went)
 brought the changes into master:  5ff40823ff4c622a79eb9897c081e40412650801

So, to solve the bit rot, the changes made were...
o Change heuristic offset from 3 to 7 to compensate for this (I think 2 px
were added top and down).
o Remove the margins from the .ui file (3 px and 2 px respectively), they
are no longer needed.
o Make the strip maximum size go from 56 px to 60 px to compensate for this
margin change.
o Code style: Add the m_ prefix to private members in the class.