Added notes sibebar and put a general layout for balloons
put the notes sidebar to the right of the view and added a
generic balloon to it to work out the formating. I changed
the background colors or the widgets to keep track of the
different widgets. This will be more usefull when the balloon
does not fill the entire sidebar.