Put the full-screen action back...
Without the fullscreen action, ctrl-h couldn't go full-screen anymore,
for a really full-screen, canvas-only view. After discussion with on irc
with Animtim, Sven and others we came to the conclusion that having both
full-screen and canvas-only is useful after all.

To avoid the confusion we had before Sven's last patch, we don't call
the canvas-only mode full-screen anymore in the gui, and show both
actions in the view menu.