After trying 2 ways of saving and parsing a PVPShape to and from an SVG, finally here is the method that is working: Save the PVPShape information as a 'rect' element in the SVG doc. It will store the transformation, size and id.
Parse the 'rect' element which has 'ViewPort' in its id as a PVPShape. Add the size, position to it.
Parse an element with tag 'calligra:frame'. Search for the shape in the list of shapes with name as the refId of this frame.
Add the animation information contained in this frame element to the shape.

When the PVPShape is parsed, it won't be painted as a rectangle at all. It has its own paint method!
:)