KoTextEditor: add convenience function to retrieve the editor
KoTextEditor::getEditorFromCanvas takes the given canvas and looks in
the set of selected shapes to see whether one of those shapes has a
KoTextShapeDataBase as userdata. If so, we return the associated
KoTextEditor.

This method is part of the KoTextEditor api because KoCanvasBase
doesn't know anything about text editors.

This is part of the plan to make Calligra not abuse the TextTool's
selection() method to get a pointer to a KoTextEditor.