SVM: Fix painting of polypolygons
To do this, we introduce a new function in the backend, specialized
for polypolygons.

Unfortunately this also means that we had to disable the so called
"complex polygons" feature when the PolyPolygon action had version >= 2.
We were never really able to understand the ideas behind them and we
could not decipher the LibreOffice source code.

Anyway, now the Desktop Summit template shows correctly in Stage.