use a special shape factory to load embedded svg images
Added a method createShapeFromOdf to KoShapeFactoryBase to create
shapes from a specific odf element. The default implementaion just
does a create the default shape and uses that to load the content
from the odf element.
Implemented a shape factory which supports loading an image from
a draw:frame element to catch embedded svg files before the image or
vector shape.
An application can choose to support creating svg shapes from an
embedded svg image instead of just displaying it as an image by
adding that shape factory to the shape registry.