Made possible to block updates only in the scheduler
This is needed for some paintops for accessing the projection of
the image. Well, ideally, they should create an additional
exclusive job for it, but sometimes the access is too short
and it is not worth creating a separate job for it. So we just added
block/unblockUpdates() calls to achieve this.