Added KisRecursiveMergeWalker
This is a composition of a full refresh walker followed with
a usual merge walker. It should be used for recomposing
subgraphs of the stack.

This looks not very good due to "diamond" inheritance. But this
is caused but using inheritance instead of a composition in the
very beginning. Now it works and all the ugliness is hidden inside
KisRecursiveMergeWalker class.

Btw, KisRecursiveMergeWalker and KisFullRefreshWalker should swap
their names to be more understandable.