fix painting of paragraph decorations for multi-line text fragments
While iterating over the lines of a multi-line text fragment to paint
the decorations line by line, the start position of the area we paint
wouldn't increase. If you had a fragment starting after the beginning
of a line, for all lines, the start position for painting decorations
would be there, instead of at the beginning of the next line.

CCBUG: 264471