pptx: Updated calculation of vertical paragraph margins.
MS PowerPoint ususes a specific approach to calculate paragraph margins when
paragraph spacing is provided as a percentage value.  Using the same approach
as in the PPT filter.

We are now compatible with the ODP output filter from MS PowerPoint2007.
The caclulated values equal to ceiled values from the MS PowerPoint2007 UI.

CCBUG:276188