refactor D&D by introducing KoView::addImages
In order to share as much code as possible, handle the basic
dragEnter and drop events in KoView, generate a list of QImage's which
are passed to the addImages method applications can implement. So Words
can insert the images anchored into the text and other apps can insert
shapes in their own preferred way.