Add cstrunner: run cstester in multiple processes
cstrunner is a tool to run cstester on multiple processes at the same time. It will create a report of the files
that where terminated by a signal. This can either happen if there is a bug in the calligra or the allowed amount of
time or cpu has been exceeded. It also creates a file md5.txt in each output directory which contains the ms5sums of
the generated thumbnails.

cstrunner uses the scripts cstwrapper and cstmd5gen

- cstwrapper
limits the resources cstester is allowed to use and makes sure the correct exit code is returned to cstrunner

- cstmd5gen
generates md5sums for a given thumbnail directory and stores them in the file md5.txt in the directory.