Change the way of painting markers
Use a QPainterPathStroker to create the outline of the path + the markers and
then fill it with the brush of the Pen. This way we get the markers and the path
to have the correct colors even when they are transparent.

This breaks the displaying of the markers in the selector as it used the outline
to paint the markers.