DOC: Updated application of CHPs for list styles.
- Starting with CHPs of the current paragraph. Those representent the difference
  from the referenced built-in text style.  GETTING THE PROPER FONT SIZE!
- Then any list related overrides are applied.

NOTE: This is a WORKAROUND!  I'm trying to provide the font-size in text-properties of the list-level-style-* element.
The proper solution in case the font-size is not provided, would be to use the font-size specified in text-properties
of the paragraph, or text-properties of it's parent.  In case of ppt/pptx files, the font-size of the first text
chunk must be used.  This complicates the filters a lot*, while keeping the layout code simple.  But OOo and the ODF output
filters do not provide the font-size information.  To stay compatible, the layout has to work differently for ODT and
for ODP files anyhow.  Let's reconsider the current approach.

* For each bullet/number with a different font-size a new list style and a new list has to be created.
  Numbered lists have to be connected to retain the proper numbering.  The worst thing is that built-in
  list styles make no sense any more.