Add marker collection and supporting parts
The KoMarkerCollection holds all markers that can be used in a document.
It should contain some predefined markers as the ones loaded in a document
if it adds additional ones (not implmented at the moment).
Add the KoMarkerSharedLoadingData which contains a lookup from the marker
name used in styles to the actual marker.
Initialize the marker collection in the KoPathShapeFactory and add to the
document resource manager