Fix a deadlock in KisTileData-copy constructor
The memory allocated by the pooler for clones is not counted
by the store in memoryHardLimit. The pooler has it's own slice of
memory and keeps track of the its size itself. So we should be able
to disable the memory check while cloning the tile data to avoid
deadlock