Fix TestKoTextEditor test
We need to check whether the rdf document is actually set in the
resourcemanager.

I will replace the current static casting horror that's all over the
place with a dedicated method after the KoResourceManager split into
KoCanvasResourceManager and KoDocumentResourceManager is done. (I
did already start on that, but it's a huge task).