Add load style properties from multiple properties.
For embedded text shapes paragraph style properties can be saved in style:paragraph-properties
or in graphic:paragraph-properties. It is also possible that some values are in one and others
are in the other e.g. in parent styles.

This patch add the possibilty to load the properties from more then one property type.

This partly fixes bug 282617. It is not fully fixed as not all properties are used by the text
shape yet.

fo:padding-top and fo:padding-bottom seem to be not yet supported.

CCBUG: 282617
REVIEW: 102737