Fixed a "dead tiles" race condition
This race condition could happen due to the fact that history writing
was based on shared pointers lifetime. That is when a tile was deleted
from a hash table, this action could be left unregistered, in case
someone held the same shared pointer. So from now on the unlink action
is registered by the hash table explicitly.

It can probably fix the hole-on-canvas bug.
CCBUG:273908