KoToolManager: only switch tools between tablet devices
This solves two bugs:

* On Ubuntu, the xinput system seems to be broken. We are getting tablet
events with unknown device/unknown pointer. If we assume that this means
that we're actually getting mouse events, everything works fine.

* Users prefer to be able to use the mouse sometimes to set the values
for the tools and options, and then use the stylus again to paint. So
do not switch back to mouse from tablet device, ever.

We now also save on activating a single-shot timer a lot of times.

BUG:285501
BUG:283130