Fix 'whole column' and absolute definitions in Excel formulas. Excel allows to define a whole column in a cell-region like "A:B" whereas that is not possible in ODF. So we need to translate that to something like "A$1:B$65536" to be sure the proper cells are evaluated in Tables.
This fixes the VLOOKUP problem visible in the document attached to bug 284601 but was a more general issue not related to only one formula but to all Excel formulas that where using such 'whole column' definitions.

Also it should fix a few cases where we did not proper evaluate absolute vs relative column/row definitions.

QA done with cstester and unittests.

BUG:284601