Implement data-value-labels in the charting plugin.
We now support reading and proper displaying of the data-label-number and data-label-text ODF elements.

This addresses bug 268481 and proper displays the values now. There is one remaining bug, inside of KDChart, which makes the pie to small if the value labels are rather large what is the case with the document attached to that report. Will work on a fix for that issue and then close bug 268481.