Fix text-shrinking logic for pie-charts in KDChart.
The bug was that while calculating with the center we
then applied the logic to calc the different to the
whole size rather and just doubled the distance. That
could lead to text being shrinked way to much like
happened at bugs.kde.org bug 268481.

This finally solves the bug (now I only need more
testing and a merge to master to mark it FIXED).
CCBUG:268481