Fixed lines appearing on a canvas on non-integer zooms
This happens because for scaling we may need some areas outside
current update rect. And these external areas may not be prepared yet
by other threads. So our current thread reads empty (not-yet-updated)
areas and uses them for scaling. This causes while lines.

Now the problem is solved like it is solved in
KisLayer::needRect/changeRect. The update, updates it's own rect, and
small adjoining areas around which depent on it, that may be called
"changeRect". Then the patch (KisImagePatch) requests even more pixels
to make the scaling right (that may be called "needRect").

BUG:286985