Make sure the brushes are also loaded as soon as possible
Loading is done in a thread, so the app isn't waiting for it, until we
try to show the brushs selector in a widget, i.e., on creating a view.
Starting the loading when we load the paintops plugin means the user
doesn't even need to notice, since he's busy doing stuff in the startup
dialog.