rdf: make KoInlineObject a QObject
The rdf system has a class that links the bookmarks (inline objects) in
the document to the rdf data: KoTextInlineRdf. This class has pointers
to KoBookmark, KoTextMeta and QTextDocument.

It is possible that bookmarks get deleted, while the rdf data still
exists. This patch wraps the pointers in QWeakPointer so we can detect
when an inline object has been deleted. To make that possible, I make
KoInlineObject inherit QObject.