Improve layout of option widgets.  Previously the option widgets would be spread out over the modebox or  tool options docker.  Before we couldn't just fix the height because then the tool option docker  tabbed with other dockers would limit the other tabs height.
Solution is adding a QWidget named SpecialSpacer that basically acts as any
other spacer but we can now detect if it is present and act accordingly.

As an option widget designer you should add such a SpecialSpacer at the bottom
of your optionwidget. Unless the widget itself can make use of expanding height
in which case you shouldn't add the SpecialSpacer.