fix loading of autofilter ranges when the first sheet does not contain any references/autofilters.
The sheet index in an Area3d token is not an index in the full list of sheets in the documents, but an index in the internal externalsheet records, so properly take that into account when parsing such a token for an auto-filter range. Slightly harder a similar fix is probably also needed for Area3D tokens in other formulas though.