Show the happy bunny!
Krita's use of KoStore is pretty old-fashioned, and we don't need the
name expansion used by the unfinished, 10-year old filter-in-filter
feature that presumably is the reason for doing weird things to paths
in a store of they start with a digit.

KoStore needs a thorough clean-up, but for now, setting disallowNameExpansion
is enough.

BUG:289242