Fixed a crash on systems where Qt is compiled with asserts
KoProgressUpdater is working in GUI thread.
ProgressHelper -- in stroke's thread

That means that the latter one has no right removing the former one.
It must be done in deferred way.

The crash itself happened in the destructor when something had to be
sent to the parent of the KoProgressUpdater.

BUG:287197
CCBUG:289602