Make update of the canvas on sigImageSizeChanged in two stages as well
Now it is done like we process usual update signals. The first part of the
signal is DirectConnection, so the canvas has an opportunity to fetch
the data of the image's projection and the second part is executed in
the context of UI using AuoConnection.

It allows us to solve a half of a really serious problem. Now we don't
need to use BlockingQueuedConnection for sigImageSizeChanged. After
fixing the second user of blocked connections -- KisNodeModel, we will be
able to fix bug 289790 quite easily by simply barrier-locking the image
before processing. Now we get a deadlock because of blocking signals.

CCBUG:289790