Fix saving of fo:color
This fixes the problem that if the outer style had style:use-window-font-color set
and the inner did not set it but used the same color for displaying text as defined
in the outer style.
Just setting the value od style:use-window-font-color="false" works in calligra but
does not work when you load the file generated by calligra in OO/LO. Therefore setting
the fo:color to the value in the properties solves the problem.

This fixes bug 290298 - font color lost on roundtrip

BUG: 290298