Fix loading the profile for my new monitor...
It seems that sometimes, cmsOpenProfileFromFile can load a profile
that cannot be loaded when we open the file itself, load all the bytes
into memory and then call cmsOpenProfileFromMem.

I verified that the reason we cannot load the profile is in there, and
not in checking whether the profile is suitable for output.