Implement a listener/facade/map structure for managing element references
It feels like Java... Or like Krita 1.3... This is what is being done
here:

* Calligra has two (for now) places where we use a particular string
throughout the lifetime of the application to cross-reference data:

 * changetracking
 * RDF

If we try to save a changetracked, rdf-containing document, both parts
will want to save an xml-id on the same element. So both parts need to
be updated on saving, when we create a new xml-id.

So, during saving, KoTextWriter will keep track of all the parts that
want to save an xml id. It will ask those parts for an implementation
of KoElementRefenceUpdateFacade and keep a list of those.

When the element is closed, a new KoElementReference will be generated
and used to save the xml-id on the element.

Of course, this means that you can only save the rdf document and the
changes _after_ saving the main text...