expose KisSelectionComponent and KisPaintDevice, not KisPixelSelection
From the KisSelection API.

Now, if you want to paint on a pixel selection component get the
selectionPaintDevice. This will be either ALPHA8 or GrayA8 -- if
the former, painting with tools won't work.

If you want to check selectedness, selected rect and so on, get
the pixelSelection. This returns a KisSelectionComponent, like
the shapeSelection() method. The relevant api has been moved to
KisSelectionComponent, which now clearly shows which part we use
for shape and which part for pixel selections.

If you want to check the selection mask, get the projections, which
is always ALPHA8.