Fix deletion of anchored object.
When deleting the shape that is part of a KoTextAnchor we also need to delete the
KoTextAnchor from the document, so that it works correctly. This adds a new method
to KoShapeBasedDocumentBase that makes it possible to attach commands that needs to
be done after a shape has been deleted.