Finished porting KisShapeController to two-stage signals processing
Yay! Now it doesn't depend on multithreading of the KisImage at all!
The updates of the internal shapes representation are done in two stage
way like it is done for all the signals of KisCanvas2. First connection
is direct and fetches the data from the image and the second is automatic
and executed in the context of UI thread.

I introduced four new signals for the image:
    void sigNodeAddedAsync(KisNodeSP node);
    void sigNodeMovedAsync(KisNodeSP node);
    void sigRemoveNodeAsync(KisNodeSP node);
    void sigLayersChangedAsync();

They are supposed to supersede old 7 signals in the nearest future.

This is the biggest part of refactoring needed for closing bug 290708.
Now the only thing left is to port KisNodeModel to use dummies stored
inside KisShapesController.

CCBUG:290708