Use the same context on saving headers
This is needed as otherwise e.g. the created image names will be the same and
as the data centers for the temporary context are not saved the images in header/footer
suddenly are the same as the onces in the documents.
This fixes bug 291117 - images in header/footer change on saving

BUG: 291117