Don't crash if sheet or area name never ends
in case there was a ' in a formula that was not closed we did not recognice
the end of the formula and crashed. Check if we are at the end and write out
a Unkonwn token and set the parse error.
Add a unit that would have cached the problem.

Thanks to Sebasian for reporting the problem.

CCMAIL: sebsauer@kdab.com