Fix bookmark-ref loading, saving and usage
This fixes bug 290302 - bookmark-ref loaded but not working nor saved

When loading bookmark-ref use the same mechanism as the toc to link internally.
When saving an internal link use bookmark-ref instead of text:a.
Make sure to keep the anchor properties when applying a style. I had a very stange lockup
of my pc when I applied the style without this fix.
Also remove the style properties from the auto style as they are not part of the saved style
data.

BUG: 290302