Introduce a Page Layout tool
The tool is supposed to handle page setup, headers, footers etc.

This is based on initial work by Jignesh.

For now it doesn't do more than move the functionality from the
format menu to it's option widgets. But it does place the
functionality in the modebox, which is important so users will
not have to relearn how to do pagelayout in a future version.

Plans are for this tool to be for page layout as the paragraph
tool was for paragraphs.