Add KoXmlReader::firstChildElement
The function works like firstChild but iterates over the child elements
and returns the first element.
Use the function in TestOpenDocumentStyle so that it works as before changing the KoXmlDocument
to no longer strip whitespace.
Rename the test name so that the TextXmlReader is executed again.
Fix TextXmlReader so it works again when whitespaces are not stripped.