Remove the old selectionChanged signal as it seems useless, replacing it with two new signals: selectionChanged() and selected()
For our docker we want a signal everytime a selection is selected. We want to be able to reapply
the paragraph style, as reapplying also removes any char styles appliend in the blocks.

We also introduce a new selectionChanged signal that is only emitted when a new selection is made
it doesn't care if the style has been modified.