Implement undo of Style changes
Two new methods are introduced: beginEdit() and endEdit() to mark
when the editing happens. At endEdit() the collected changes
are applied to any attached documents like before.

The difference is that it now happens through ChangeStylesMacroCommand
Which in turn uses ChangeStylesCommand

The role of ChangeStylesMacroCommand is to act as a macro plus to
make sure the styles within the stylemanager itself are updated.

The role of ChangeStylesCommand is to update the qtextdocuments.

REVIEW: 103779