Update the imagecollection map when necessary
The ImageCollection contains a map of ImageData objects where the key
is generated from the contents of the image. If setImage is called, the
key is updated inside the ImageData object, but not inside the map.

This leads to problems when we start saving and try to find the right image
data objects in the collection based on the key stored in the documents.

This patch makes sure that the image collection's list of keys is kept
up to date.