Fixed managing selection masks which are children of the root node
Now the mapping between KisNodeDummy and QModelIndex is done by
classes KisModeIndexConverter and KisModelIndexConverterShowAll.
They represent two options available to the user: normal mode and
mode when the root layer is visible.

This is the last commit in a series of refactorings for KisNodeModel.
Now there should be no known crashes or bugs. So if you see one, do
not hesitate to report! ;)

My plan is the following:
1) I'm pushing it to my branch.
2) Then I'll fix drag-and-drop (it is unrelated to the refactorings)
3) Then I'll make a call for testing both of the features, to
   save the time of artists ;)

BUG:291271
BUG:290708