Improve TOS apply text styles of graphic style
This loads the text styles that are part of the graphic styles of a shape and apply it
in the text shape used for TOS.
This patch also contains a workaount for the ppt/pptx filter to ignore padding in case
there is a custom shape used. This keeps basically keeps the state we had before when the
properties where not applied. This is needed as the TOS does not yet support all that is
needed to keep the quality of the rendered documents high. Thanks to Matus Uzak for this
part of the patch.

BUG: 295158