Fix breakage in of display of master shapes.
Remove stuff that was committed in the xml:id patch but does not belong there.
Looks like a merge error.

This fixes the problem that the shapes of the master page are shown even if they
should not be displayed.

Can someone please review so I can backport.

CCMAIL: calligra-devel@kde.org