Fix saving of shapes in the correct order.
The saving of the order is currently broken for quite a lot of files.

One example is slide 4 of

http://www.zagge.de/anchor/bug/open.umich.edu%252Fsites%252Fdefault%252Ffiles%252FSeverance-SI5%25302-W%25309-Week11-Sesons-and-Cookies.odp

This was caused by the commit 3fd6c20f31f946f3128af63fc433449f36c06c9d that saves the
z-index out to the file. However the way it is done is not fully correct. I works mostly
for words. With this patch we go to back to the original behaviour of not saving the z-index
as it is defined by the order of the shapes in the xml file. We still save the z-index for
words.

Please review.

I have run cstester --roundtip and it improves for odp and not regressions for odt files.

CCMAIL: calligra-devel@kde.org

Reviewed-by: boemann
(cherry picked from commit 5af08715b5f9e8fb16ef9a9c48e3507203e40c1e)