Fix roundtrip testing for use-windows-font-color
In case use-windows-font-color is used don't set the hardcoded font to black
as this overwrites the use-windows-font-color and therefore it is no longer
used after loading the file again.

This is visible in quite a lot of our documents. Without that fix all these docs
have a different pixels set which is not visible for the user but is shown in
visualimagecomapre.

This is e.g. visible in the file:
assets.en.oreilly.com%2F1%2Fevent%2F21%2FMySQL%20and%20Search%20at%20Craigslist%20Presentation.odp

Can you please review so I can commit to the 2.4 branch.

Tested with cstester --roundtrip and it makes cstester usable again when the flag is used.

CCMAIL: calliga-devel@kde.org

Reviewed-by: boemann
(cherry picked from commit d222d758d0b8a0d38249bc391b5649a42d6b184a)