libmsooxml: parse and use line styles from the DrawingML theme properly
TODO: At the moment the ln element is read before lnRef, which
defines a reference to the line style which properties are
overwritten by the ln element.  The approach to inherit only
missing properties has quite good results so this can be fixed
later.

----

I finally stumbled across the problem that a subset of methods
defined in MsooXmlCommonReaderDrawingMLMethods.h MUST be shared
with readers of DrawinML tables and themes.  The methods defined
threin are meant to be included into filters as a whole.

I have put the initial subset of read_ methods into a new pair of
header files.  There's still a lot of work to remove the duplicit
implementations from mentioned DrawingML readers.