Fix loading of docx files
On Windows, these pesky temporary files cannot be closed and reopened,
everything dies if they are around and we try to create a new file
around them. So delete the object that keeps a file handle open on them.

This fixes loading DOCX files, but not yet DOC files. And we're leaving
dead temporary files behind us.