libmso: Added arrow heads support.
* The basic implementation from PPT filter was moved into libmso
and MarkerStyles synced with LO.

* Disabled MarkerStyle creation when the default graphic style is
requested by the client, it's shape specific so it does NOT
belong there.

* DOC: Disabled processing of the OfficeArtDggContainer when
shapes are beeing processed by libmso.  Use default property
values instead of document defaults (sync with the PPT filter).

TODO: The size is defined by an enumeration value and seems to be
arrow type specific.